Add SPDX identifiers
The Software Package Data Exchange[0] (SPDX) is a Linux Foundation
initiative to make it easier for people to see exactly what licenses the
software they use come under.

For this we simply add a single line at the top of each source file with
the appropriate header, in this case

    /* SPDX-License-Identifier: LGPL-2.1 */

As part of this we rename the COPYING (LGPL license text) file to
LGPL-2.1 (to match the SPDX identifier).

We also then remove some redundant text about the license in each file.

[0] - https://spdx.org/

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>